# docs-site/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Arka.fund Catalog and NAV API",
    "version": "1.0.0",
    "description": "Public, read-only indexed data for Arka discovery, NAV, assets, managers, activity and operational monitoring. Exact monetary values are returned as integer strings to preserve on-chain precision.",
    "contact": {
      "name": "Arka.fund",
      "url": "https://arka.fund/"
    },
    "license": {
      "name": "Public API terms",
      "url": "https://arka.fund/docs/"
    }
  },
  "servers": [
    {
      "url": "https://catalog.arka.fund",
      "description": "Stellar mainnet"
    }
  ],
...
  "paths": {
    "/v1/nav": {
      "get": {
        "tags": [
          "NAV"
        ],
        "summary": "Read aggregate NAV",
        "operationId": "getCatalogNav",
        "description": "Returns the canonical snapshot-backed NAV aggregate, denomination totals, valuation state and indexer monitoring state. Use the activity endpoints for live contract events.",
        "responses": {
          "200": {
            "description": "Current aggregate NAV response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NavResponse"
                }
              }
            }
          },
          "503": {
            "description": "The latest indexed snapshot is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/arkas": {
      "get": {
        "tags": [
          "Arkas"
        ],
        "summary": "List indexed Arkas",
        "operationId": "listArkas",
        "description": "Returns paginated Arkas. Use curated=true and delisted=false for the public platform listing.",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "description": "Arka sort field.",
            "schema": {
              "type": "string",
              "enum": [
                "nav",
                "manager",
                "syncedAt"
              ]
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "description": "Sort direction.",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "curated",
            "in": "query",
            "required": false,
            "description": "Filter by platform curation state.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "delisted",
            "in": "query",
            "required": false,
            "description": "Filter by delisting state.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Case-insensitive address or identity search.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Zero-based result offset.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of returned items.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated Arka list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArkaPage"
                }
              }
            }
          }
        }
      }
    },
    "/v1/arkas/{id}": {
      "get": {
        "tags": [
          "Arkas"
        ],
        "summary": "Read an Arka",
        "operationId": "getArka",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Arka contract ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Indexed Arka state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Arka"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "The latest indexed snapshot is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/arkas/{id}/identity": {
      "get": {
        "tags": [
          "Arkas"
        ],
        "summary": "Read an Arka profile",
        "operationId": "getArkaIdentity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Arka contract ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public Arka profile metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArkaIdentity"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Arkas"
        ],
        "summary": "Update an Arka profile",
        "operationId": "updateArkaIdentity",
        "description": "Saves public Arka profile metadata after verifying a Stellar signature from the current Arka manager.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Arka contract ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated Arka profile metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArkaIdentity"
                }
              }
            }
          },
          "400": {
            "description": "The signed profile payload is invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The signer is not authorized to update this profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/arkas/{id}/assets": {
      "get": {
        "tags": [
          "Arkas"
        ],
        "summary": "List Arka asset exposures",
        "operationId": "listArkaAssets",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Arka contract ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Arka asset exposures.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetExposure"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "The latest indexed snapshot is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/arkas/{id}/portfolio": {
      "get": {
        "tags": [
          "Arkas"
        ],
        "summary": "Read ranked Arka composition",
        "operationId": "getArkaPortfolio",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Arka contract ID.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of positions.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked portfolio composition.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "The latest indexed snapshot is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/arkas/{id}/history": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Read Arka NAV history",
        "operationId": "getArkaHistory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Arka contract ID.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "description": "Inclusive ISO-8601 start time.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "description": "Inclusive ISO-8601 end time.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "description": "Sort direction.",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of history points.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 365
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated Arka history.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/v1/arkas/{id}/activity": {
      "get": {
        "tags": [
          "Activity"
        ],
        "summary": "Read Arka contract activity",
        "operationId": "getArkaActivity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Arka contract ID.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "description": "Activity type.",
            "schema": {
              "type": "string",
              "enum": [
                "deposit",
                "redeem",
                "profit",
                "lend",
                "borrow",
                "repay",
                "withdraw"
              ]
            }
          },
          {
            "name": "fromLedger",
            "in": "query",
            "required": false,
            "description": "Minimum ledger sequence.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "toLedger",
            "in": "query",
            "required": false,
            "description": "Maximum ledger sequence.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "description": "Sort direction.",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of activity records.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated activity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityPage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ArkaPage": {
        "type": "object",
        "required": [
          "total",
          "offset",
          "limit",
          "items"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Arka"
            }
          },
          "dataStatus": {
            "type": "string",
            "enum": [
              "live",
              "unavailable"
            ]
          },
          "unavailableReason": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}